refactor: DocumentServiceTest 중복 삭제, ApplicationSourceDocumentRepository 파일 위치 수정
